🚚 Move the API from /api/private to /api/v1 (RAL-1261)
Renames the public API route ahead of opening it to all Pro users.

- Hono basePath and route directory move to /api/v1; docs at /api/v1/docs,
  spec at /api/v1/openapi
- OpenAPI document becomes "Rallly API" v1.0.0 with a versioning note
  (v1 stable, breaking changes only via a new version prefix)
- Temporary alias route at /api/private/[...route] keeps beta integrations
  working (~60 days). A next.config rewrite can't do this because route
  handlers see the original request URL, which Hono would not match
- Wide-event service and rate limiter key renamed private-api → api
