+++ apps/web/src/app/api/v1/utils/api-key.ts
