+++ apps/web/src/app/api/v1/utils/time-slots.ts
