+++ apps/web/src/features/api-keys/constants.ts
@@ -10,10 +10,10 @@ export const API_RATE_LIMIT_PER_MINUTE = 60;
 
 /**
  * Path segment for the current version of the public API. Keeping it here
- * means moving between versions (e.g. `private` → `v1`) only touches one
+ * means moving between versions (e.g. `v1` → `v2`) only touches one
  * place, and every link stays pointed at the latest docs.
  */
-const CURRENT_API_VERSION_PATH = "private";
+const CURRENT_API_VERSION_PATH = "v1";
 
 /**
  * Link to the interactive API reference for the current version. Callers
